Move samples and refactor them for simplification
