+++ automl/cloud-client/src/main/java/com/example/automl/GetOperationStatus.java
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package com.google.cloud.translate.automl;
+package com.example.automl;
 
-// [START automl_translate_get_operation_status]
+// [START automl_get_operation_status]
 import com.google.cloud.automl.v1.AutoMlClient;
 import com.google.longrunning.Operation;
 
@@ -54,4 +54,4 @@ class GetOperationStatus {
     }
   }
 }
-// [END automl_translate_get_operation_status]
+// [END automl_get_operation_status]
