+++ automl/cloud-client/src/main/java/com/example/automl/ImportDataset.java
@@ -14,9 +14,9 @@
  * limitations under the License.
  */
 
-package com.google.cloud.translate.automl;
+package com.example.automl;
 
-// [START automl_translate_import_dataset]
+// [START automl_import_dataset]
 import com.google.cloud.automl.v1.AutoMlClient;
 import com.google.cloud.automl.v1.DatasetName;
 import com.google.cloud.automl.v1.GcsSource;
@@ -57,4 +57,4 @@ class ImportDataset {
     }
   }
 }
-// [END automl_translate_import_dataset]
+// [END automl_import_dataset]
