+++ automl/cloud-client/src/test/java/com/example/automl/TranslatePredictIT.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 Google Inc.
+ * Copyright 2019 Google LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.translate.automl;
+package com.example.automl;
 
 import static com.google.common.truth.Truth.assertThat;
 
@@ -27,10 +27,10 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
-/** Tests for translation "PredictionAPI" sample. */
+/** Tests for translation "Predict" sample. */
 @RunWith(JUnit4.class)
 @SuppressWarnings("checkstyle:abbreviationaswordinname")
-public class PredictionApiIT {
+public class TranslatePredictIT {
   private static final String COMPUTE_REGION = "us-central1";
   private static final String PROJECT_ID = "java-docs-samples-testing";
   private static final String modelId = "TRL2188848820815848149";
@@ -51,9 +51,9 @@ public class PredictionApiIT {
   }
 
   @Test
-  public void testPredict() throws Exception {
+  public void testPredict() {
     // Act
-    PredictionApi.predict(PROJECT_ID, COMPUTE_REGION, modelId, filePath);
+    TranslatePredict.predict(PROJECT_ID, modelId, filePath);
 
     // Assert
     String got = bout.toString();
